requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
msedge-selenium-tools==3.141.2
PyYAML==5.3.1
requests==2.25.0
selenium==3.141.0
urllib3==1.25.10
